+++ organizeit/src/main/java/com/organizeit/db/repository/ItemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ItemRepository extends CrudRepository<Item, Integer> {
-    List<Item> findItemsByDrawerId(int id);
+public interface ItemRepository extends CrudRepository<Item, Long> {
+    List<Item> findItemsByDrawerId(long id);
 }
