+++ organizeit/src/main/java/com/organizeit/db/repository/ShelfListRepository.java
@@ -3,6 +3,6 @@ package com.organizeit.db.repository;
 import com.organizeit.db.entity.ShelfList;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShelfListRepository extends CrudRepository<ShelfList, Integer> {
+public interface ShelfListRepository extends CrudRepository<ShelfList, Long> {
 
 }
